+++ CHPlugin/Source/Extensions/CRToast+Extensions.swift
@@ -14,11 +14,15 @@ import SwiftyJSON
 extension CRToastManager {
   static func showErrorFromData(_ data: Data?) {
     guard data != nil else { return }
-    let json = JSON(data: data!)
-    guard let errors: [CHErrorResponse] = Mapper<CHErrorResponse>()
-      .mapArray(JSONObject: json["errors"].object) else { return }
-    errors.forEach { (error) -> () in
-      CRToastManager.showErrorMessage(error.message)
+    do {
+      let json = try JSON(data: data!)
+      guard let errors: [CHErrorResponse] = Mapper<CHErrorResponse>()
+        .mapArray(JSONObject: json["errors"].object) else { return }
+      errors.forEach { (error) -> () in
+        CRToastManager.showErrorMessage(error.message)
+      }
+    } catch {
+      return
     }
   }