+++ CHPlugin/Source/Utils/CHUtils.swift
@@ -71,7 +71,7 @@ class CHUtils {
     if let path = bundle.path(forResource: "countryInfo", ofType: "json") {
       do {
         let data = try Data(contentsOf: URL(fileURLWithPath: path), options: .alwaysMapped)
-        let jsonObj = JSON(data: data)
+        let jsonObj = try JSON(data: data)
         if jsonObj != JSON.null {
           return Mapper<CHCountry>()
             .mapArray(JSONObject: jsonObj.object)